Move FriendRequestButtons to separate file [PAR-48] [PAR-49]
